+++ governance-app/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Support for update timeout parameters transaction.
 - Support for update minimum block time transaction.
 - Support for update block energy limit transaction.
+- Support for update finalization committee parameters.
 
 ## 0.0.1